Copy kernel-source (openwrt not build in thie version)

+++ openwrt/install.sh
@@ -29,6 +29,10 @@ if [ ! -e feeds ]; then
   mkdir feeds;
 fi
 
+if [ ! -e kernel ]; then
+  mkdir kernel;
+fi
+
 . ./config/config.common
 
 if [ "x$CONFIG" != "x" ]; then
@@ -38,9 +42,9 @@ if [ "x$CONFIG" != "x" ]; then
 fi
 
 
-for i in $EXTRA_FEEDS; do
-  (cd feeds; git clone $i)
-done
+#for i in $EXTRA_FEEDS; do
+#  (cd feeds; git clone $i)
+#done
 
 FEED_PATHS=`(cd feeds; ls)`
 
@@ -48,6 +52,8 @@ for i in $BUILD_ARCH; do
   . config/config.$i
 
   OPENWRT_PATH=$OPENWRT_VERSION\-$i
+
+if [ "x$DIR" = "y" ]; then
   (git clone $OPENWRT_GITURL $OPENWRT_PATH)
   (cd $OPENWRT_PATH; git reset --hard $OPENWRT_REVISION)
 
@@ -76,6 +82,18 @@ for i in $BUILD_ARCH; do
 
   echo "export PATH=\$PATH:\$TOOLCHAIN_DIR/bin/" >> $OPENWRT_PATH/toolchain.bashrc
   echo "export PATH=\$PATH:\$STAGING_DIR/host/bin" >> $OPENWRT_PATH/toolchain.bashrc
+
+fi
+
+  if [ "x$KERNEL_CONFIG" != "x" ]; then
+
+    BUILD_DIR=`cd $OPENWRT_PATH; ls -d build_dir/target-*`
+    LINUXTOOL_DIR=`cd $OPENWRT_PATH/$BUILD_DIR; ls -d linux-*`
+    LINUX_DIR=`cd $OPENWRT_PATH/$BUILD_DIR/$LINUXTOOL_DIR; ls -d linux-3*`
+
+    cp -r $OPENWRT_PATH/$BUILD_DIR/$LINUXTOOL_DIR/$LINUX_DIR kernel/$LINUX_DIR\-$i
+
+  fi
 
 done
